# aspnetcore/src/IO.Swagger/Controllers/ProfileControllerApi.cs
/*
 * ProfileController
 *
 * Handles profile detail.
 *
 * OpenAPI spec version: 1.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class ProfileControllerApiController : ControllerBase
    { 
        /// <summary>
        /// Complete Driver Profile
        /// </summary>
        /// <remarks>Complete Driver profile after registration</remarks>
        /// <param name="body"></param>
        /// <response code="200">OK</response>
        [HttpPost]
        [Route("/profile/Completedriverprofile")]
        [ValidateModelState]
        [SwaggerOperation("3")]
        [SwaggerResponse(statusCode: 200, type: typeof(string), description: "OK")]
        public virtual IActionResult 3([FromBody]CompleteDriverProfileDto body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(string));
            string exampleJson = null;
            exampleJson = "\"\"";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<string>(exampleJson)
                        : default(string);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Admin gets the detail of Driver</remarks>
        /// <param name="auth">Token</param>
        /// <param name="body"></param>
        /// <response code="200">OK</response>
        [HttpPost]
        [Route("/profile/GetDriverProfile")]
        [ValidateModelState]
        [SwaggerOperation("4")]
        [SwaggerResponse(statusCode: 200, type: typeof(CompleteDriverProfileDto), description: "OK")]
        public virtual IActionResult 4([FromHeader][Required()]string auth, [FromBody]Body body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(CompleteDriverProfileDto));
            string exampleJson = null;
            exampleJson = "{\n  \"Experience\" : \"Experience\",\n  \"InsuranceDocument\" : [ \"\", \"\" ],\n  \"Address\" : \"Address\",\n  \"WorkPhone\" : \"WorkPhone\",\n  \"HomePhone\" : \"HomePhone\",\n  \"PreferredPhone\" : \"PreferredPhone\",\n  \"VehicleRegisteration\" : [ \"\", \"\" ],\n  \"AvailabilityDays\" : [ \"\", \"\" ],\n  \"DriverLicense\" : [ \"\", \"\" ],\n  \"MobilePhone\" : \"MobilePhone\",\n  \"Age\" : \"Age\",\n  \"ProfilePicture\" : [ \"\", \"\" ]\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<CompleteDriverProfileDto>(exampleJson)
                        : default(CompleteDriverProfileDto);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}